+++ cmd/operator-sdk/migrate/cmd.go
@@ -19,19 +19,17 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/ansible"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/helm"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/input"
+	"github.com/operator-framework/operator-sdk/internal/scaffold"
+	"github.com/operator-framework/operator-sdk/internal/scaffold/ansible"
+	"github.com/operator-framework/operator-sdk/internal/scaffold/helm"
+	"github.com/operator-framework/operator-sdk/internal/scaffold/input"
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
 
-	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
 var (
-	depManager string
 	headerFile string
 	repo       string
 )
@@ -45,9 +43,8 @@ func NewCmd() *cobra.Command {
 		RunE:  migrateRun,
 	}
 
-	newCmd.Flags().StringVar(&depManager, "dep-manager", "modules", `Dependency manager the new project will use (choices: "dep", "modules")`)
 	newCmd.Flags().StringVar(&headerFile, "header-file", "", "Path to file containing headers for generated Go files. Copied to hack/boilerplate.go.txt")
-	newCmd.Flags().StringVar(&repo, "repo", "", "Project repository path. Used as the project's Go import path. This must be set if outside of $GOPATH/src with Go modules, and cannot be set if --dep-manager=dep")
+	newCmd.Flags().StringVar(&repo, "repo", "", "Project repository path. Used as the project's Go import path. This must be set if outside of $GOPATH/src (e.g. github.com/example-inc/my-operator)")
 
 	return newCmd
 }
@@ -76,7 +73,7 @@ func migrateRun(cmd *cobra.Command, args []string) error {
 }
 
 func verifyFlags() error {
-	err := projutil.CheckDepManagerWithRepo(projutil.DepManagerType(depManager), repo)
+	err := projutil.CheckRepo(repo)
 	if err != nil {
 		return err
 	}
@@ -119,11 +116,9 @@ func migrateAnsible() error {
 		s.BoilerplatePath = headerFile
 	}
 
-	if err := scaffoldAnsibleDepManager(s, cfg); err != nil {
-		return errors.Wrap(err, "migrate Ansible dependency manager file scaffold failed")
-	}
-
 	err = s.Execute(cfg,
+		&ansible.GoMod{},
+		&scaffold.Tools{},
 		&ansible.Main{},
 		&dockerfile,
 		&ansible.Entrypoint{},
@@ -160,11 +155,9 @@ func migrateHelm() error {
 		s.BoilerplatePath = headerFile
 	}
 
-	if err := scaffoldHelmDepManager(s, cfg); err != nil {
-		return errors.Wrap(err, "migrate Helm dependency manager file scaffold failed")
-	}
-
 	err := s.Execute(cfg,
+		&helm.GoMod{},
+		&scaffold.Tools{},
 		&helm.Main{},
 		&helm.DockerfileHybrid{
 			Watches:    true,
@@ -188,30 +181,4 @@ func renameDockerfile() error {
 	}
 	log.Infof("Renamed Dockerfile to %s and replaced with newer version. Compare the new Dockerfile to your old one and manually migrate any customizations", newDockerfilePath)
 	return nil
-}
-
-func scaffoldHelmDepManager(s *scaffold.Scaffold, cfg *input.Config) error {
-	var files []input.File
-	switch m := projutil.DepManagerType(depManager); m {
-	case projutil.DepManagerDep:
-		files = append(files, &helm.GopkgToml{})
-	case projutil.DepManagerGoMod:
-		files = append(files, &helm.GoMod{}, &scaffold.Tools{})
-	default:
-		return projutil.ErrInvalidDepManager(depManager)
-	}
-	return s.Execute(cfg, files...)
-}
-
-func scaffoldAnsibleDepManager(s *scaffold.Scaffold, cfg *input.Config) error {
-	var files []input.File
-	switch m := projutil.DepManagerType(depManager); m {
-	case projutil.DepManagerDep:
-		files = append(files, &ansible.GopkgToml{})
-	case projutil.DepManagerGoMod:
-		files = append(files, &ansible.GoMod{}, &scaffold.Tools{})
-	default:
-		return projutil.ErrInvalidDepManager(depManager)
-	}
-	return s.Execute(cfg, files...)
 }