+++ cmd/operator-sdk/new/cmd.go
@@ -22,10 +22,10 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/ansible"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/helm"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/input"
+	"github.com/operator-framework/operator-sdk/internal/scaffold"
+	"github.com/operator-framework/operator-sdk/internal/scaffold/ansible"
+	"github.com/operator-framework/operator-sdk/internal/scaffold/helm"
+	"github.com/operator-framework/operator-sdk/internal/scaffold/input"
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
 
 	"github.com/pkg/errors"
@@ -56,11 +56,10 @@ generates a skeletal app-operator application in $HOME/projects/example.com/app-
 	newCmd.Flags().StringVar(&apiVersion, "api-version", "", "Kubernetes apiVersion and has a format of $GROUP_NAME/$VERSION (e.g app.example.com/v1alpha1) - used with \"ansible\" or \"helm\" types")
 	newCmd.Flags().StringVar(&kind, "kind", "", "Kubernetes CustomResourceDefintion kind. (e.g AppService) - used with \"ansible\" or \"helm\" types")
 	newCmd.Flags().StringVar(&operatorType, "type", "go", "Type of operator to initialize (choices: \"go\", \"ansible\" or \"helm\")")
-	newCmd.Flags().StringVar(&depManager, "dep-manager", "modules", `Dependency manager the new project will use (choices: "dep", "modules")`)
-	newCmd.Flags().StringVar(&repo, "repo", "", "Project repository path for Go operators. Used as the project's Go import path. This must be set if outside of $GOPATH/src with Go modules, and cannot be set if --dep-manager=dep (e.g. github.com/example-inc/my-opertor)")
+	newCmd.Flags().StringVar(&repo, "repo", "", "Project repository path for Go operators. Used as the project's Go import path. This must be set if outside of $GOPATH/src (e.g. github.com/example-inc/my-operator)")
 	newCmd.Flags().BoolVar(&gitInit, "git-init", false, "Initialize the project directory as a git repository (default false)")
 	newCmd.Flags().StringVar(&headerFile, "header-file", "", "Path to file containing headers for generated Go files. Copied to hack/boilerplate.go.txt")
-	newCmd.Flags().BoolVar(&makeVendor, "vendor", false, "Use a vendor directory for dependencies. This flag only applies when --dep-manager=modules (the default)")
+	newCmd.Flags().BoolVar(&makeVendor, "vendor", false, "Use a vendor directory for dependencies")
 	newCmd.Flags().BoolVar(&skipValidation, "skip-validation", false, "Do not validate the resulting project's structure and dependencies. (Only used for --type go)")
 	newCmd.Flags().BoolVar(&generatePlaybook, "generate-playbook", false, "Generate a playbook skeleton. (Only used for --type ansible)")
 
@@ -76,7 +75,6 @@ var (
 	kind             string
 	operatorType     string
 	projectName      string
-	depManager       string
 	headerFile       string
 	repo             string
 	gitInit          bool
@@ -180,26 +178,13 @@ func doGoScaffold() error {
 		s.BoilerplatePath = headerFile
 	}
 
-	var err error
-	switch m := projutil.DepManagerType(depManager); m {
-	case projutil.DepManagerDep:
-		err = s.Execute(cfg, &scaffold.GopkgToml{})
-	case projutil.DepManagerGoMod:
-		if goModOn, merr := projutil.GoModOn(); merr != nil {
-			return merr
-		} else if !goModOn {
-			return errors.New(`dependency manager "modules" requires working directory to be in $GOPATH/src` +
-				` and GO111MODULE=on, or outside of $GOPATH/src and GO111MODULE="on", "auto", or unset. More info: https://github.com/operator-framework/operator-sdk/blob/master/doc/user-guide.md#go-modules`)
-		}
-		err = s.Execute(cfg, &scaffold.GoMod{}, &scaffold.Tools{})
-	default:
-		err = projutil.ErrNoDepManager
-	}
-	if err != nil {
-		return fmt.Errorf("dependency manager file scaffold failed: (%v)", err)
+	if err := projutil.CheckGoModules(); err != nil {
+		return err
 	}
 
-	err = s.Execute(cfg,
+	err := s.Execute(cfg,
+		&scaffold.GoMod{},
+		&scaffold.Tools{},
 		&scaffold.Cmd{},
 		&scaffold.Dockerfile{},
 		&scaffold.Entrypoint{},
@@ -384,12 +369,7 @@ func verifyFlags() error {
 			return fmt.Errorf("operators of type Go do not use --api-version or --kind")
 		}
 
-		dm := projutil.DepManagerType(depManager)
-		if !makeVendor && dm == projutil.DepManagerDep {
-			log.Warnf("--dep-manager=dep requires a vendor directory; ignoring --vendor=false")
-		}
-		err := projutil.CheckDepManagerWithRepo(dm, repo)
-		if err != nil {
+		if err := projutil.CheckRepo(repo); err != nil {
 			return err
 		}
 	}
@@ -424,30 +404,20 @@ func execProjCmd(cmd string, args ...string) error {
 }
 
 func getDeps() error {
-	switch m := projutil.DepManagerType(depManager); m {
-	case projutil.DepManagerDep:
-		log.Info("Running dep ensure")
-		if err := execProjCmd("dep", "ensure", "-v"); err != nil {
-			return err
-		}
-	case projutil.DepManagerGoMod:
-		// Only when a user requests a vendor directory be created should
-		// "go mod vendor" be run during project initialization.
-		if makeVendor {
-			log.Info("Running go mod vendor")
-			opts := projutil.GoCmdOptions{
-				Args: []string{"-v"},
-				Dir:  filepath.Join(projutil.MustGetwd(), projectName),
-			}
-			if err := projutil.GoCmd("mod vendor", opts); err != nil {
-				return err
-			}
-		} else {
-			// Avoid done message.
-			return nil
-		}
-	default:
-		return projutil.ErrInvalidDepManager(depManager)
+
+	// Only when a user requests a vendor directory be created should
+	// "go mod vendor" be run during project initialization.
+	if !makeVendor {
+		return nil
+	}
+
+	log.Info("Running go mod vendor")
+	opts := projutil.GoCmdOptions{
+		Args: []string{"-v"},
+		Dir:  filepath.Join(projutil.MustGetwd(), projectName),
+	}
+	if err := projutil.GoCmd("mod vendor", opts); err != nil {
+		return err
 	}
 	log.Info("Done getting dependencies")
 	return nil
@@ -463,24 +433,20 @@ func initGit() error {
 }
 
 func validateProject() error {
-	switch projutil.DepManagerType(depManager) {
-	case projutil.DepManagerGoMod:
-		log.Info("Validating project")
-		// Run "go build ./..." to make sure all packages can be built
-		// correctly. From "go help build":
-		//
-		//	When compiling multiple packages or a single non-main package,
-		//	build compiles the packages but discards the resulting object,
-		//	serving only as a check that the packages can be built.
-		opts := projutil.GoCmdOptions{
-			PackagePath: "./...",
-			Dir:         filepath.Join(projutil.MustGetwd(), projectName),
-		}
-		if err := projutil.GoBuild(opts); err != nil {
-			return err
-		}
-		log.Info("Project validation successful.")
+	log.Info("Validating project")
+	// Run "go build ./..." to make sure all packages can be built
+	// correctly. From "go help build":
+	//
+	//	When compiling multiple packages or a single non-main package,
+	//	build compiles the packages but discards the resulting object,
+	//	serving only as a check that the packages can be built.
+	opts := projutil.GoCmdOptions{
+		PackagePath: "./...",
+		Dir:         filepath.Join(projutil.MustGetwd(), projectName),
 	}
-
+	if err := projutil.GoBuild(opts); err != nil {
+		return err
+	}
+	log.Info("Project validation successful.")
 	return nil
 }