+++ cmd/operator-sdk/scorecard/cmd.go
@@ -17,8 +17,9 @@ package scorecard
 import (
 	"fmt"
 
-	"github.com/operator-framework/operator-sdk/internal/pkg/scorecard"
-	scplugins "github.com/operator-framework/operator-sdk/internal/pkg/scorecard/plugins"
+	"github.com/operator-framework/operator-sdk/internal/scorecard"
+	schelpers "github.com/operator-framework/operator-sdk/internal/scorecard/helpers"
+	scplugins "github.com/operator-framework/operator-sdk/internal/scorecard/plugins"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -36,12 +37,14 @@ func NewCmd() *cobra.Command {
 	scorecardCmd.Flags().String(scorecard.ConfigOpt, "", fmt.Sprintf("config file (default is '<project_dir>/%s'; the config file's extension and format can be .yaml, .json, or .toml)", scorecard.DefaultConfigFile))
 	scorecardCmd.Flags().String(scplugins.KubeconfigOpt, "", "Path to kubeconfig of custom resource created in cluster")
 	scorecardCmd.Flags().StringP(scorecard.OutputFormatOpt, "o", scorecard.TextOutputFormat, fmt.Sprintf("Output format for results. Valid values: %s, %s", scorecard.TextOutputFormat, scorecard.JSONOutputFormat))
+	scorecardCmd.Flags().String(schelpers.VersionOpt, schelpers.DefaultScorecardVersion, fmt.Sprintf("scorecard version (tech preview version is '%s'", schelpers.LatestScorecardVersion))
 
 	// TODO: make config file global and make this a top level flag
 	viper.BindPFlag(scorecard.ConfigOpt, scorecardCmd.Flags().Lookup(scorecard.ConfigOpt))
 
 	viper.BindPFlag("scorecard."+scplugins.KubeconfigOpt, scorecardCmd.Flags().Lookup(scplugins.KubeconfigOpt))
 	viper.BindPFlag("scorecard."+scorecard.OutputFormatOpt, scorecardCmd.Flags().Lookup(scorecard.OutputFormatOpt))
+	viper.BindPFlag("scorecard."+schelpers.VersionOpt, scorecardCmd.Flags().Lookup(schelpers.VersionOpt))
 
 	return scorecardCmd
 }