+++ hack/image/build-scorecard-proxy-image.sh
@@ -2,6 +2,8 @@
 
 set -eux
 
+source hack/lib/image_lib.sh
+
 # build operator binary and base image
 WD="$(dirname "$(pwd)")"
 GOOS=linux GOARCH=amd64 CGO_ENABLED=0 \
@@ -12,4 +14,6 @@ GOOS=linux GOARCH=amd64 CGO_ENABLED=0 \
   images/scorecard-proxy/cmd/proxy/main.go
 pushd images/scorecard-proxy
 docker build -t "$1" .
+# If using a kind cluster, load the image into all nodes.
+load_image_if_kind "$1"
 popd