+++ pkg/kube-metrics/server.go
@@ -19,7 +19,7 @@ import (
 	"net"
 	"net/http"
 
-	kcollector "k8s.io/kube-state-metrics/pkg/collector"
+	metricsstore "k8s.io/kube-state-metrics/pkg/metrics_store"
 )
 
 const (
@@ -27,11 +27,11 @@ const (
 	healthzPath = "/healthz"
 )
 
-func ServeMetrics(collectors [][]kcollector.Collector, host string, port int32) {
+func ServeMetrics(stores [][]*metricsstore.MetricsStore, host string, port int32) {
 	listenAddress := net.JoinHostPort(host, fmt.Sprint(port))
 	mux := http.NewServeMux()
 	// Add metricsPath
-	mux.Handle(metricsPath, &metricHandler{collectors})
+	mux.Handle(metricsPath, &metricHandler{stores})
 	// Add healthzPath
 	mux.HandleFunc(healthzPath, func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(200)
@@ -55,7 +55,7 @@ func ServeMetrics(collectors [][]kcollector.Collector, host string, port int32) 
 }
 
 type metricHandler struct {
-	collectors [][]kcollector.Collector
+	stores [][]*metricsstore.MetricsStore
 }
 
 func (m *metricHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
@@ -63,9 +63,9 @@ func (m *metricHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// 0.0.4 is the exposition format version of prometheus
 	// https://prometheus.io/docs/instrumenting/exposition_formats/#text-based-format
 	resHeader.Set("Content-Type", `text/plain; version=`+"0.0.4")
-	for _, collectors := range m.collectors {
-		for _, c := range collectors {
-			c.Collect(w)
+	for _, stores := range m.stores {
+		for _, s := range stores {
+			s.WriteAll(w)
 		}
 	}
 }